Update model to spec

+++ server/src/controllers/questions.ts
@@ -13,18 +13,20 @@ export const getQuestionById = (id: models.classes.guid) => {
 
 // Add question to the list
 export const addQuestion = (question: models.dtos.QuestionDto) => {
-	return questionsRepository.addItem(question);
+	return questionsRepository.addItem({ ...question, lastUpdate: Date.now(), active: false });
 };
 
 export const editQuestion = async (
 	id: models.classes.guid,
 	updatedQuestion: models.dtos.QuestionDto
 ) => {
-	await questionsRepository.updateItem(id, updatedQuestion);
+	await questionsRepository.updateItem(id, { ...updatedQuestion, lastUpdate: Date.now() });
 	//update old question ==> handled by Repo
 	//push  edited question to db ==> handled by Repo
 };
 
 export const deleteQuestion = async (id: models.classes.guid) => {
+	const question = await getQuestionById(id);
+	if (question.active) { throw new BadRequestError("This question cannot be deleted!"); }
 	await questionsRepository.deleteItem(id);
 };
